+++ Source/Scene/GlobeDepth.js
@@ -169,6 +169,7 @@ define([
         if (!defined(globeDepth._clearColorCommand)) {
             globeDepth._clearColorCommand = new ClearCommand({
                 color : new Color(0.0, 0.0, 0.0, 0.0),
+                stencil : 0.0,
                 owner : globeDepth
             });
         }